# use O(N) build-heap in vqsort