The previous build-heap algorithm is O(N*logN).

Reference: *Introduction to Algorithms*, third edition, chap 6.3