@QuarticCat 2022-06-08
My CPU is Ryzen 3700X.

Before this commit:

```text
[ RUN      ] BenchSortGroup/BenchSort.BenchAllSort/AVX2
      AVX2:          std:     f32: uniform32: 1.00E+06   59 MB/s ( 1 threads)
      AVX2:         heap:     f32: uniform32: 1.00E+06   25 MB/s ( 1 threads)
      AVX2:           vq:     f32: uniform32: 1.00E+06  791 MB/s ( 1 threads)
      AVX2:          std:     i32: uniform32: 1.00E+06   71 MB/s ( 1 threads)
      AVX2:         heap:     i32: uniform32: 1.00E+06   30 MB/s ( 1 threads)
      AVX2:           vq:     i32: uniform32: 1.00E+06  753 MB/s ( 1 threads)
      AVX2:          std:     i64: uniform32: 1.00E+06  142 MB/s ( 1 threads)
      AVX2:         heap:     i64: uniform32: 1.00E+06   60 MB/s ( 1 threads)
      AVX2:           vq:     i64: uniform32: 1.00E+06  757 MB/s ( 1 threads)
      AVX2:          std:    u128: uniform32: 1.00E+06  147 MB/s ( 1 threads)
      AVX2:         heap:    u128: uniform32: 1.00E+06   60 MB/s ( 1 threads)
      AVX2:           vq:    u128: uniform32: 1.00E+06  708 MB/s ( 1 threads)
[       OK ] BenchSortGroup/BenchSort.BenchAllSort/AVX2 (25606 ms)
```

After this commit:

```text
[ RUN      ] BenchSortGroup/BenchSort.BenchAllSort/AVX2
      AVX2:          std:     f32: uniform32: 1.00E+06   61 MB/s ( 1 threads)
      AVX2:         heap:     f32: uniform32: 1.00E+06   26 MB/s ( 1 threads)
      AVX2:           vq:     f32: uniform32: 1.00E+06  830 MB/s ( 1 threads)
      AVX2:          std:     i32: uniform32: 1.00E+06   72 MB/s ( 1 threads)
      AVX2:         heap:     i32: uniform32: 1.00E+06   32 MB/s ( 1 threads)
      AVX2:           vq:     i32: uniform32: 1.00E+06  836 MB/s ( 1 threads)
      AVX2:          std:     i64: uniform32: 1.00E+06  148 MB/s ( 1 threads)
      AVX2:         heap:     i64: uniform32: 1.00E+06   62 MB/s ( 1 threads)
      AVX2:           vq:     i64: uniform32: 1.00E+06  759 MB/s ( 1 threads)
      AVX2:          std:    u128: uniform32: 1.00E+06  149 MB/s ( 1 threads)
      AVX2:         heap:    u128: uniform32: 1.00E+06   62 MB/s ( 1 threads)
      AVX2:           vq:    u128: uniform32: 1.00E+06  713 MB/s ( 1 threads)
[       OK ] BenchSortGroup/BenchSort.BenchAllSort/AVX2 (24349 ms)
```

Improvement is not quite significant and varies on my machine. But generally the O(N) build-heap version is slightly faster.