@jan-wassenberg 2022-06-08
Thanks @QuarticCat for the patch! Would you mind adding some comments to the SiftDown function?

I notice that the second run seems to be faster across the board. Let's check whether there may be some throttling going on. If CPU temp hits 90, I believe AMD will throttle. We can verify that with `perf stat -e cycles,ref-cycles bench_sort`.

Do we know the CPU temps? Lower = less leakage power, which may indirectly allow higher frequencies (by enabling higher voltage without reaching the power cap).