@QuarticCat 2022-06-08
The `SiftDown` function is basically copied from the original `HeapSort`'s sift-down part. I just modified some variable names.