@QuarticCat 2022-06-08
It seems that `ref-cycles` is not supported on my machine.

From `sensors`, my CPU temp is floating between 76°C and 79°C during the benchmark.