@QuarticCat 2022-06-17
To make the result more accurate, I tried the following steps:

1. Fix CPU frequency

```console
$ sudo cpupower frequency-set -u 3G -d 3G
```

From `watch -n 0 rg \"cpu MHz\" /proc/cpuinfo` I can observe that the CPU frequencies are fixed to `3600 MHz` (not 3G, I don't know why).

However, I can still see some cores are boosted to 4.2 GHz during the benchmark.

2. Disable boost

```console
$ sudo sh -c 'echo 0 > /sys/devices/system/cpu/cpufreq/boost'
```

Now the boost disappears. But the benchmark results are still floating.

3. Warm up and run multiple times

First I disabled most benches. Below is what it looks like now.

```text
> bazel-bin/hwy/contrib/sort/bench_sort
Running main() from gmock_main.cc
[==========] Running 4 tests from 1 test suite.
[----------] Global test environment set-up.
[----------] 4 tests from BenchSortGroup/BenchSort
[ RUN      ] BenchSortGroup/BenchSort.BenchAllSort/AVX2
      AVX2:         heap:     i32: uniform32: 1.00E+06   27 MB/s ( 1 threads)
      AVX2:         heap:     i64: uniform32: 1.00E+06   52 MB/s ( 1 threads)
[       OK ] BenchSortGroup/BenchSort.BenchAllSort/AVX2 (9227 ms)
[ RUN      ] BenchSortGroup/BenchSort.BenchAllSort/SSE4
[       OK ] BenchSortGroup/BenchSort.BenchAllSort/SSE4 (0 ms)
[ RUN      ] BenchSortGroup/BenchSort.BenchAllSort/SSSE3
[       OK ] BenchSortGroup/BenchSort.BenchAllSort/SSSE3 (0 ms)
[ RUN      ] BenchSortGroup/BenchSort.BenchAllSort/Emu128
[       OK ] BenchSortGroup/BenchSort.BenchAllSort/Emu128 (0 ms)
[----------] 4 tests from BenchSortGroup/BenchSort (9227 ms total)

[----------] Global test environment tear-down
[==========] 4 tests from 1 test suite ran. (9227 ms total)
[  PASSED  ] 4 tests.
```

Then I use hyperfine to run this benchmark multiple times.

```console
$ hyperfine --warmup=2 --runs=10 'bazel-bin/hwy/contrib/sort/bench_sort'
```

Here are the final results.

Before this commit:

```text
Benchmark 1: bazel-bin/hwy/contrib/sort/bench_sort
  Time (mean ± σ):      9.502 s ±  0.020 s    [User: 9.491 s, System: 0.002 s]
  Range (min … max):    9.477 s …  9.545 s    10 runs
```

After this commit:

```text
Benchmark 1: bazel-bin/hwy/contrib/sort/bench_sort
  Time (mean ± σ):      9.267 s ±  0.018 s    [User: 9.256 s, System: 0.003 s]
  Range (min … max):    9.240 s …  9.308 s    10 runs
```

I am completely a beginner at benchmarking. You can pull this PR and bench it by yourself.

If you accept this PR, I will solve conflicts a few hours later. Now I need to sleep for a while.