@jan-wassenberg 2022-06-21
@QuarticCat nice, looks like you have done more extensive benchmarking than is often the case, and it looks like a solid win :) We're happy to accept this PR after you sync it to HEAD.